integrated BeamNG.ai executor
